Test for redirect and handling pax_global_header
FetchTarball currently uses an http client that has a redirect
policy to automatically redirect. In the future this should
be injected with a client that has a custom redirection policy

Added a test to check that redirects are taken care of in
FetchTarball.

Also, pax_global_header was a problem. Dealing with it in
DefaultExtractor.