Added the first cut of Angmar
Angmar is a struct that takes a queue client,
an extractor generator and an api client. It currently
uses a single worker to fetch the urls provided into a
named queue that the queue client can pop from.

New packages were created to deal with queue client
and createServer was moved to testutils from gh_test

main has now changed to use Angmar.