Made Angmar run concurrent jobs with workers
There are now concurrent workers that run jobs on Angmar.
The structure of Angmar message has been slightly modified as well.
As has the structure of the job. A little clunky right now. Should
get better.

GithubAPI is now a proper DownloadClient. GithubAPI now takes
a value receiver instead of a pointer receiver on DownloadClient