Stripping source mount point from basepath of gen
We had to assume that Angmar will take the responsibility of
placing the right message on Uruk queues. In order to do so
it has to remove the source mount point location of the source
volume and place a relative path on the queue.

It does this by stripping the source mount point from the
generator's base path